write queries for file tree, metrics and findings and add a port/usecase/adapter/controller for quering file content and metrics

+++ coderadar-core/src/main/java/io/reflectoring/coderadar/analyzer/service/AnalyzeCommitService.java
@@ -114,8 +114,12 @@ public class AnalyzeCommitService implements AnalyzeCommitUseCase {
                 finding.getCharEnd()));
       }
       metricValues.add(
-          new MetricValue(
-              metric.getId(), fileMetrics.getMetricCount(metric), commitId, fileId, findings));
+          new MetricValue( // TODO: intValue()...
+              metric.getId(),
+              fileMetrics.getMetricCount(metric).intValue(),
+              commitId,
+              fileId,
+              findings));
     }
     return metricValues;
   }
